+++ core/network/src/test/java/dev/jvoyatz/modern/android/network/moshi/MoshiTest.kt
@@ -1,15 +1,13 @@
-package dev.jvoyatz.modern.android.network.config
+package dev.jvoyatz.modern.android.network.moshi
 
 import com.google.common.truth.Truth
-import com.squareup.moshi.Json
 import com.squareup.moshi.JsonDataException
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dev.jvoyatz.modern.android.network.config.ApiResponseCallAdapterFactory
 import dev.jvoyatz.modern.android.network.config.model.ApiResponse
 import dev.jvoyatz.modern.android.network.utils.getResFileContent
 import dev.jvoyatz.modern.android.network.utils.isTypeOf
-import kotlinx.coroutines.Deferred
-import kotlinx.coroutines.runBlocking
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.After
@@ -17,84 +15,13 @@ import org.junit.Before
 import org.junit.Test
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-import retrofit2.http.GET
-
-
-//models
-internal data class BoredActivity(
-    @Json(name = "activity")
-    val activity: String
-)
-
-internal data class BoredActivityInvalid(
-    @Json(name = "activityii")
-    val activity: String
-)
-
-internal data class BoredActivityError(
-    @Json(name = "error")
-    val error: String
-)
-
-internal data class BoredActivityErrorInvalid(
-    @Json(name = "errorrrr")
-    val error: String
-)
-
-//api
-
-internal interface BoredApi {
-    @GET("/activity")
-    suspend fun getRandomActivity(): ApiResponse<BoredActivity, BoredActivityError>
-
-    @GET("/activity")
-    suspend fun getRandomActivityInvalid(): ApiResponse<BoredActivityInvalid, BoredActivityErrorInvalid>
-
-    @GET("/activity")
-    fun getDeferredRandomActivity(): Deferred<ApiResponse<BoredActivity, BoredActivityError>>
-
-    @GET("/activity")
-    fun getDeferredRandomActivityInvalid(): Deferred<ApiResponse<BoredActivityInvalid, BoredActivityErrorInvalid>>
-
-    @GET("/activity")
-    suspend fun getTestUnit(): ApiResponse<Unit, BoredActivityError>
-
-    @GET("/activity")
-    fun getDeferredTestUnit(): Deferred<ApiResponse<Unit, BoredActivityErrorInvalid>>
-}
-
-internal class BoredApiDataSource(
-    private val api: BoredApi
-) {
-    fun getRandomActivity() = runBlocking {
-        api.getRandomActivity()
-    }
-
-    fun getRandomActivityInvalid() = runBlocking {
-        api.getRandomActivityInvalid()
-    }
-
-    fun getDeferredRandomActivity() = runBlocking {
-        api.getDeferredRandomActivity().await()
-    }
-    fun getDeferredRandomActivityInvalid() = runBlocking {
-        api.getDeferredRandomActivityInvalid().await()
-    }
-
-    fun getTestUnit() = runBlocking {
-        api.getTestUnit()
-    }
-
-    fun getDeferredTestUnit() = runBlocking {
-        api.getDeferredTestUnit().await()
-    }
-}
 
 class MoshiTest {
     private lateinit var server: MockWebServer
     private lateinit var sut: BoredApiDataSource
 
     private val filename = "/bored_api_activity_sample.json"
+
     @Before
     fun setup() {
         server = MockWebServer()
@@ -134,7 +61,8 @@ class MoshiTest {
         val response = sut.getRandomActivity()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<BoredActivity, BoredActivityError>>()).isTrue()
+        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<BoredActivity, BoredActivityError>>())
+            .isTrue()
         val resp = response as ApiResponse.ApiSuccess<BoredActivity, BoredActivityError>
         Truth.assertThat(resp.body.activity).isEqualTo(expected)
     }
@@ -151,7 +79,8 @@ class MoshiTest {
         val response = sut.getRandomActivity()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.HttpError<BoredActivity, BoredActivityError>>()).isTrue()
+        Truth.assertThat(response.isTypeOf<ApiResponse.HttpError<BoredActivity, BoredActivityError>>())
+            .isTrue()
         val resp = response as ApiResponse.HttpError<BoredActivity, BoredActivityError>
         Truth.assertThat(resp.errorBody?.error).isEqualTo(expected)
     }
@@ -167,8 +96,10 @@ class MoshiTest {
         val response = sut.getRandomActivityInvalid()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>()).isTrue()
-        val resp = response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
+        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>())
+            .isTrue()
+        val resp =
+            response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
         Truth.assertThat(resp.error?.isTypeOf<JsonDataException>()).isTrue()
     }
 
@@ -178,16 +109,20 @@ class MoshiTest {
         val expected = "Think of a new business idea"
         server.enqueue(
             MockResponse().setResponseCode(404)
-                .setBody("""
+                .setBody(
+                    """
                     { "test": "down" }
-                """.trimIndent())
+                """.trimIndent()
+                )
         )
         //when
         val response = sut.getRandomActivityInvalid()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>()).isTrue()
-        val resp = response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
+        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>())
+            .isTrue()
+        val resp =
+            response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
         Truth.assertThat(resp.error?.isTypeOf<JsonDataException>()).isTrue()
     }
 
@@ -202,8 +137,10 @@ class MoshiTest {
         val response = sut.getRandomActivityInvalid()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>()).isTrue()
-        val resp = response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
+        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>())
+            .isTrue()
+        val resp =
+            response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
         Truth.assertThat(resp.error?.isTypeOf<JsonDataException>()).isTrue()
     }
 
@@ -219,7 +156,8 @@ class MoshiTest {
         val response = sut.getDeferredRandomActivity()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<BoredActivity, BoredActivityError>>()).isTrue()
+        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<BoredActivity, BoredActivityError>>())
+            .isTrue()
         val resp = response as ApiResponse.ApiSuccess<BoredActivity, BoredActivityError>
         Truth.assertThat(resp.body.activity).isEqualTo(expected)
     }
@@ -236,7 +174,8 @@ class MoshiTest {
         val response = sut.getRandomActivity()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.HttpError<BoredActivity, BoredActivityError>>()).isTrue()
+        Truth.assertThat(response.isTypeOf<ApiResponse.HttpError<BoredActivity, BoredActivityError>>())
+            .isTrue()
         val resp = response as ApiResponse.HttpError<BoredActivity, BoredActivityError>
         Truth.assertThat(resp.errorBody?.error).isEqualTo(expected)
     }
@@ -252,8 +191,10 @@ class MoshiTest {
         val response = sut.getDeferredRandomActivityInvalid()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>()).isTrue()
-        val resp = response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
+        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>())
+            .isTrue()
+        val resp =
+            response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
         Truth.assertThat(resp.error?.isTypeOf<JsonDataException>()).isTrue()
     }
 
@@ -268,8 +209,10 @@ class MoshiTest {
         val response = sut.getDeferredRandomActivityInvalid()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>()).isTrue()
-        val resp = response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
+        Truth.assertThat(response.isTypeOf<ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>>())
+            .isTrue()
+        val resp =
+            response as ApiResponse.UnexpectedError<BoredActivityInvalid, BoredActivityErrorInvalid>
         Truth.assertThat(resp.error?.isTypeOf<JsonDataException>()).isTrue()
     }
 
@@ -283,7 +226,8 @@ class MoshiTest {
         val response = sut.getTestUnit()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<Unit, BoredActivityError>>()).isTrue()
+        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<Unit, BoredActivityError>>())
+            .isTrue()
     }
 
     @Test
@@ -296,6 +240,7 @@ class MoshiTest {
         val response = sut.getDeferredTestUnit()
 
         //then
-        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<Unit, BoredActivityError>>()).isTrue()
+        Truth.assertThat(response.isTypeOf<ApiResponse.ApiSuccess<Unit, BoredActivityError>>())
+            .isTrue()
     }
 }
